--- libs/category/src/models/category-applied-filter.ts
@@ -1,10 +1,0 @@
-import { DaffCategoryFilterBase } from './category-filter-base';
-
-export interface DaffCategoryAppliedFilter extends DaffCategoryFilterBase {
-	options: DaffCategoryAppliedFilterOption[];
-}
-
-export interface DaffCategoryAppliedFilterOption {
-	label: string;
-	value: string;
-}
