--- libs/category/src/models/category-filter-base.ts
@@ -1,7 +1,0 @@
-import { DaffCategoryFilterType } from './filters/public_api';
-
-export interface DaffCategoryFilterBase {
-	label: string;
-	name: string;
-	type: DaffCategoryFilterType;
-}
