--- libs/category/src/models/category-filter.ts
@@ -1,10 +1,0 @@
-import { DaffCategoryAppliedFilterOption } from './category-applied-filter';
-import { DaffCategoryFilterBase } from './category-filter-base';
-
-export interface DaffCategoryFilter extends DaffCategoryFilterBase {
-	options: DaffCategoryFilterOption[];
-}
-
-export interface DaffCategoryFilterOption extends DaffCategoryAppliedFilterOption {
-	count: number;
-}
