+++ libs/category/src/models/filters/filter-request.ts
@@ -1,0 +1,6 @@
+import { DaffCategoryFilterEqualRequest } from './type/equal/request';
+import { DaffCategoryFilterRangeNumericRequest } from './type/range-numeric/request';
+
+export type DaffCategoryFilterRequest =
+	DaffCategoryFilterEqualRequest |
+	DaffCategoryFilterRangeNumericRequest;
