+++ libs/category/src/models/filters/filter-toggle-request.ts
@@ -1,0 +1,6 @@
+import { DaffCategoryFilterEqualToggleRequest } from './type/equal/toggle-request';
+import { DaffCategoryFilterRangeNumericToggleRequest } from './type/range-numeric/public_api';
+
+export type DaffCategoryFilterToggleRequest =
+	DaffCategoryFilterEqualToggleRequest |
+	DaffCategoryFilterRangeNumericToggleRequest;
