+++ libs/category/src/models/filters/public_api.ts
@@ -1,3 +1,9 @@
+export { DaffCategoryFilterBase } from './category-filter-base';
+export { DaffCategoryFilter } from './category-filter';
+export { DaffCategoryFilterOption } from './category-filter-option';
+export { DaffCategoryFilterRequest } from './filter-request';
+export { DaffCategoryFilterToggleRequest } from './filter-toggle-request';
+export { DaffCategoryFilterable } from './category-filterable';
 export { DaffCategoryFilterType } from './type/category-filter-type';
 
 export * from './type/equal/public_api';
