+++ libs/category/src/models/filters/type/equal/public_api.ts
@@ -1,3 +1,4 @@
 export { DaffCategoryFilterEqual } from './filter';
 export { DaffCategoryFilterEqualOption } from './option';
-
+export { DaffCategoryFilterEqualRequest } from './request';
+export { DaffCategoryFilterEqualToggleRequest } from './toggle-request';
