+++ libs/category/src/models/filters/type/equal/request.ts
@@ -1,0 +1,7 @@
+import { DaffCategoryFilterType } from '../category-filter-type';
+
+export interface DaffCategoryFilterEqualRequest {
+	type: DaffCategoryFilterType.Equal;
+	name: string;
+	value: string[];
+}
