+++ libs/category/src/models/filters/type/equal/toggle-request.ts
@@ -1,0 +1,7 @@
+import { DaffCategoryFilterType } from '../category-filter-type';
+
+export interface DaffCategoryFilterEqualToggleRequest {
+	type: DaffCategoryFilterType.Equal;
+	name: string;
+	value: string;
+}
