+++ libs/category/src/models/filters/type/range-numeric/public_api.ts
@@ -1,1 +1,3 @@
 export { DaffCategoryFilterRangeNumeric } from './filter';
+export { DaffCategoryFilterRangeNumericRequest } from './request';
+export { DaffCategoryFilterRangeNumericToggleRequest } from './toggle-request';
