+++ libs/category/src/models/filters/type/range-numeric/request.ts
@@ -1,0 +1,6 @@
+import { DaffCategoryFilterType } from '../category-filter-type';
+import { DaffCategoryFilterRangeRequestBase } from '../range/public_api';
+
+export interface DaffCategoryFilterRangeNumericRequest extends DaffCategoryFilterRangeRequestBase<number> {
+	type: DaffCategoryFilterType.RangeNumeric;
+}
