+++ libs/category/src/models/filters/type/range-numeric/toggle-request.ts
@@ -1,0 +1,6 @@
+import { DaffCategoryFilterType } from '../category-filter-type';
+import { DaffCategoryFilterRangeRequestBase } from '../range/request-base';
+
+export interface DaffCategoryFilterRangeNumericToggleRequest extends DaffCategoryFilterRangeRequestBase<number> {
+	type: DaffCategoryFilterType.RangeNumeric;
+}
