+++ libs/category/src/models/filters/type/range/public_api.ts
@@ -1,6 +1,7 @@
 export { DaffCategoryFilterRangeBase } from './filter-base';
 export { DaffCategoryFilterRangeOption } from './option';
 export { DaffCategoryFilterRangePair } from './pair';
+export { DaffCategoryFromToFilterSeparator } from './separator';
 export { DaffCategoryFilterRangeRequest } from './request';
 export { DaffCategoryFilterRangeRequestBase } from './request-base';
 export { DaffCategoryFilterRangeRequestOption } from './request-option';
