+++ libs/category/src/models/filters/type/range/request-option.ts
@@ -1,0 +1,4 @@
+export interface DaffCategoryFilterRangeRequestOption<T> {
+  min: T;
+  max: T;
+}
