+++ libs/category/src/models/filters/type/range/request.ts
@@ -1,0 +1,3 @@
+import { DaffCategoryFilterRangeNumericRequest } from '../range-numeric/public_api';
+
+export type DaffCategoryFilterRangeRequest = DaffCategoryFilterRangeNumericRequest;
