+++ libs/category/src/models/filters/type/range/separator.ts
@@ -1,0 +1,5 @@
+/**
+ * The separator between a range type filter's values
+ * e.g. "price": "30-40"
+ */
+export const DaffCategoryFromToFilterSeparator = '-';
