+++ libs/category/src/models/public_api.ts
@@ -1,14 +1,7 @@
 export { DaffGetCategoryResponse } from './get-category-response';
 export * from './filters/public_api';
-export {
-  DaffCategoryFilter,
-  DaffCategoryFilterOption,
-} from './category-filter';
 export { DaffCategoryPageConfigurationState } from './category-page-configuration-state';
 export { DaffCategoryBreadcrumb } from './category-breadcrumb';
 export { DaffCategory } from './category';
 export { DaffGenericCategory } from './generic-category';
 export { DaffCategoryRequest } from './category-request';
-export * from './requests/filter-request';
-export * from './category-filter-base';
-export * from './category-applied-filter';
